feat: started writing stdlib prelude, made selective imports and fixed re-definitions in imports

+++ vio_helpers/vio_string/vio_string.h
@@ -6,9 +6,9 @@
 #define VIO_HELPERS_VIO_STRINGS_H
 
 #include <stddef.h>
-#include <stdint.h>
 #include <stdlib.h>
-#include <string.h>
+
+
 
 typedef struct {
   uint32_t ref_count;
@@ -25,6 +25,6 @@ VioString vio_str_from_literal(const char *s, size_t len);
 VioString vio_str_concat(VioString a, VioString b);
 void vio_str_retain(VioString s);
 void vio_str_release(VioString s);
-size_t vio_str_len(VioString s);
+int64_t vio_str_len(VioString s);
 
 #endif // VIO_HELPERS_VIO_STRINGS_H
